+++ examples/demo-rollup/tests/bank/zk_rollup/bank_tests.rs
@@ -4,16 +4,19 @@ use sov_bank::event::Event as BankEvent;
 use sov_bank::utils::TokenHolder;
 use sov_bank::Coins;
 use sov_cli::NodeClient;
+use sov_demo_rollup::MockDemoRollup;
 use sov_mock_da::storable::service::StorableMockDaService;
 use sov_mock_da::BlockProducingConfig;
+use sov_modules_api::execution_mode::Native;
 use sov_modules_api::OperatingMode;
 use sov_rollup_interface::node::da::DaServiceWithRetries;
 use sov_rollup_interface::node::ledger_api::FinalityStatus;
+use sov_test_utils::test_rollup::{get_appropriate_rollup_prover_config, RollupBuilder};
 use sov_test_utils::TestSpec;
 
 use crate::bank::helpers::*;
 use crate::bank::{DaLayerTxSender, SequencerTxSender, TxSender, TOKEN_NAME};
-use crate::test_helpers::*;
+use crate::test_helpers::test_genesis_paths;
 
 const BLOCK_PRODUCING_CONFIG: BlockProducingConfig = BlockProducingConfig::OnSubmit;
 
@@ -24,13 +27,14 @@ async fn bank_tx_tests_instant_finality_using_sequencer_tx_submission() -> anyho
         finalization_blocks: 0,
     };
 
-    let test_rollup = TestRollup::create_test_rollup_in_memory_da(
-        get_appropriate_rollup_prover_config(),
-        BLOCK_PRODUCING_CONFIG,
-        test_case.finalization_blocks,
-        OperatingMode::Zk,
-    )
-    .await?;
+    let test_rollup =
+        RollupBuilder::<MockDemoRollup<Native>>::start_memory_da_rollup_in_the_background(
+            get_appropriate_rollup_prover_config(),
+            BLOCK_PRODUCING_CONFIG,
+            test_case.finalization_blocks,
+            &test_genesis_paths(OperatingMode::Zk),
+        )
+        .await?;
 
     let sender = SequencerTxSender {};
 
@@ -49,13 +53,14 @@ async fn bank_tx_tests_non_instant_finality_using_sequencer_tx_submission() -> a
         wait_for_aggregated_proof: false,
         finalization_blocks: 2,
     };
-    let test_rollup = TestRollup::create_test_rollup_in_memory_da(
-        get_appropriate_rollup_prover_config(),
-        BLOCK_PRODUCING_CONFIG,
-        test_case.finalization_blocks,
-        OperatingMode::Zk,
-    )
-    .await?;
+    let test_rollup =
+        RollupBuilder::<MockDemoRollup<Native>>::start_memory_da_rollup_in_the_background(
+            get_appropriate_rollup_prover_config(),
+            BLOCK_PRODUCING_CONFIG,
+            test_case.finalization_blocks,
+            &test_genesis_paths(OperatingMode::Zk),
+        )
+        .await?;
 
     let sender = SequencerTxSender {};
 
@@ -75,13 +80,14 @@ async fn bank_tx_tests_instant_finality_using_da_layer_tx_submission() -> anyhow
         finalization_blocks: 0,
     };
 
-    let test_rollup = TestRollup::create_test_rollup_in_memory_da(
-        get_appropriate_rollup_prover_config(),
-        BLOCK_PRODUCING_CONFIG,
-        test_case.finalization_blocks,
-        OperatingMode::Zk,
-    )
-    .await?;
+    let test_rollup =
+        RollupBuilder::<MockDemoRollup<Native>>::start_memory_da_rollup_in_the_background(
+            get_appropriate_rollup_prover_config(),
+            BLOCK_PRODUCING_CONFIG,
+            test_case.finalization_blocks,
+            &test_genesis_paths(OperatingMode::Zk),
+        )
+        .await?;
 
     let sender = DaLayerTxSender::new(test_rollup.da_service.clone());
     // If the rollup throws an error, return it and stop trying to send the transaction